Refactored files into header and cpp files
